// scripts/out/open-api/typescript-node/model/attribute.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class Attribute {
    'id'?: number;
    'name'?: string;
    'imageUrl'?: string;
    'isOn'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "imageUrl",
            "baseName": "imageUrl",
            "type": "string"
        },
        {
            "name": "isOn",
            "baseName": "isOn",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return Attribute.attributeTypeMap;
    }
}