// scripts/out/open-api/typescript-node/model/attributeType.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class AttributeType {
    'id'?: number;
    'name'?: string;
    'hasYesOption'?: boolean;
    'hasNoOption'?: boolean;
    'yesIconUrl'?: string;
    'noIconUrl'?: string;
    'notChosenIconUrl'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "hasYesOption",
            "baseName": "hasYesOption",
            "type": "boolean"
        },
        {
            "name": "hasNoOption",
            "baseName": "hasNoOption",
            "type": "boolean"
        },
        {
            "name": "yesIconUrl",
            "baseName": "yesIconUrl",
            "type": "string"
        },
        {
            "name": "noIconUrl",
            "baseName": "noIconUrl",
            "type": "string"
        },
        {
            "name": "notChosenIconUrl",
            "baseName": "notChosenIconUrl",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AttributeType.attributeTypeMap;
    }
}