// scripts/out/open-api/typescript-node/model/friendRequest.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { UserReference } from './userReference';

export class FriendRequest {
    'id'?: number;
    'requestorUserCode'?: string;
    'requestor'?: UserReference;
    'requestedUserCode': string;
    'requested'?: UserReference;
    'message': string;
    'isOutgoing'?: boolean;
    'requestDateUtc'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "requestorUserCode",
            "baseName": "requestorUserCode",
            "type": "string"
        },
        {
            "name": "requestor",
            "baseName": "requestor",
            "type": "UserReference"
        },
        {
            "name": "requestedUserCode",
            "baseName": "requestedUserCode",
            "type": "string"
        },
        {
            "name": "requested",
            "baseName": "requested",
            "type": "UserReference"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "isOutgoing",
            "baseName": "isOutgoing",
            "type": "boolean"
        },
        {
            "name": "requestDateUtc",
            "baseName": "requestDateUtc",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return FriendRequest.attributeTypeMap;
    }
}