// scripts/out/open-api/typescript-node/model/geocacheCount.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GeocacheType } from './geocacheType';

export class GeocacheCount {
    'geocacheType'?: GeocacheType;
    'count'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "geocacheType",
            "baseName": "geocacheType",
            "type": "GeocacheType"
        },
        {
            "name": "count",
            "baseName": "count",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return GeocacheCount.attributeTypeMap;
    }
}