// scripts/out/open-api/typescript-node/model/geocacheLimit.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class GeocacheLimit {
    'liteCallsRemaining'?: number;
    'liteCallsSecondsToLive'?: number;
    'fullCallsRemaining'?: number;
    'fullCallsSecondsToLive'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "liteCallsRemaining",
            "baseName": "liteCallsRemaining",
            "type": "number"
        },
        {
            "name": "liteCallsSecondsToLive",
            "baseName": "liteCallsSecondsToLive",
            "type": "number"
        },
        {
            "name": "fullCallsRemaining",
            "baseName": "fullCallsRemaining",
            "type": "number"
        },
        {
            "name": "fullCallsSecondsToLive",
            "baseName": "fullCallsSecondsToLive",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return GeocacheLimit.attributeTypeMap;
    }
}