// scripts/out/open-api/typescript-node/model/geocacheLog.ts
/**
 * API v1.0
 * API version 1.0.
 *
 * The version of the OpenAPI document: 1.0.0-oas3
 * Contact: apihelp@geocaching.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Coordinates } from './coordinates';
import { GeocacheLogType } from './geocacheLogType';
import { Image } from './image';
import { User } from './user';

export class GeocacheLog {
    'referenceCode'?: string;
    'ownerCode'?: string;
    'owner'?: User;
    'imageCount'?: number;
    'isEncoded'?: boolean;
    'isArchived'?: boolean;
    'images'?: Array<Image>;
    'url'?: string;
    'geocacheName'?: string;
    'ianaTimezoneId'?: string;
    'loggedDate': Date;
    'text': string;
    'type'?: string;
    'geocacheLogType'?: GeocacheLogType;
    'updatedCoordinates'?: Coordinates;
    'geocacheCode': string;
    'usedFavoritePoint'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "referenceCode",
            "baseName": "referenceCode",
            "type": "string"
        },
        {
            "name": "ownerCode",
            "baseName": "ownerCode",
            "type": "string"
        },
        {
            "name": "owner",
            "baseName": "owner",
            "type": "User"
        },
        {
            "name": "imageCount",
            "baseName": "imageCount",
            "type": "number"
        },
        {
            "name": "isEncoded",
            "baseName": "isEncoded",
            "type": "boolean"
        },
        {
            "name": "isArchived",
            "baseName": "isArchived",
            "type": "boolean"
        },
        {
            "name": "images",
            "baseName": "images",
            "type": "Array<Image>"
        },
        {
            "name": "url",
            "baseName": "url",
            "type": "string"
        },
        {
            "name": "geocacheName",
            "baseName": "geocacheName",
            "type": "string"
        },
        {
            "name": "ianaTimezoneId",
            "baseName": "ianaTimezoneId",
            "type": "string"
        },
        {
            "name": "loggedDate",
            "baseName": "loggedDate",
            "type": "Date"
        },
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "geocacheLogType",
            "baseName": "geocacheLogType",
            "type": "GeocacheLogType"
        },
        {
            "name": "updatedCoordinates",
            "baseName": "updatedCoordinates",
            "type": "Coordinates"
        },
        {
            "name": "geocacheCode",
            "baseName": "geocacheCode",
            "type": "string"
        },
        {
            "name": "usedFavoritePoint",
            "baseName": "usedFavoritePoint",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return GeocacheLog.attributeTypeMap;
    }
}